Add files via upload
O termo "moniker" foi traduzido como ""cabeçalho expandido. Ver notas abaixo: 
A moniker is the internal representation of a routine header with

• no articles preceding parameter types,
• parameter types in square brackets, and
• expanded prepositions.
 
Each piece of a moniker is a monikette. For example:
 
Routine header: To add a number to a count:
Moniker: add [number] in/into/to [count]
Monikettes: (1) add (2) [number] (3) in/into/to (4) [count]
 
We use this internal representation because it...
 
• makes it easy reduce types (like [count] to [number]) when looking for an appropriate routine to call;
• makes it easy to expand prepositions (like "to" into "in/into/to"); and
• shows how the compiler interpreted various phrases in error messages.